[PIDM-35] feat: Refactor api iban update
