+++ src/main/java/it/gov/pagopa/debtposition/controller/payments/api/IPaymentsController.java
@@ -9,6 +9,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import it.gov.pagopa.debtposition.model.ProblemJson;
+import it.gov.pagopa.debtposition.model.payments.UpdateTransferIbanMassiveModel;
 import it.gov.pagopa.debtposition.model.payments.PaymentOptionModel;
 import it.gov.pagopa.debtposition.model.payments.response.PaymentOptionModelResponse;
 import it.gov.pagopa.debtposition.model.payments.response.PaymentOptionWithDebtorInfoModelResponse;
@@ -266,4 +267,56 @@ public interface IPaymentsController {
           @PathVariable("nav")
           String nav,
       @Valid @RequestBody NotificationFeeUpdateModel notificationFeeUpdateModel);
+
+    @Operation(
+            summary = "The Organization updates the IBANs of every unpaid payment option's transfers",
+            security = {
+                    @SecurityRequirement(name = "ApiKey"),
+                    @SecurityRequirement(name = "Authorization")
+            },
+            operationId = "updateTransferIbanMassive")
+    @ApiResponses(
+            value = {
+                    @ApiResponse(responseCode = "200", description = "IBANs updated"),
+                    @ApiResponse(
+                            responseCode = "400",
+                            description = "Malformed request.",
+                            content =
+                            @Content(
+                                    mediaType = MediaType.APPLICATION_JSON_VALUE,
+                                    schema = @Schema(implementation = ProblemJson.class))),
+                    @ApiResponse(
+                            responseCode = "401",
+                            description = "Wrong or missing function key.",
+                            content = @Content(schema = @Schema())),
+                    @ApiResponse(
+                            responseCode = "404",
+                            description = "No valid IBAN found.",
+                            content = @Content(schema = @Schema(implementation = ProblemJson.class))),
+                    @ApiResponse(
+                            responseCode = "422",
+                            description = "Unprocessable: no payment position in valid state.",
+                            content =
+                            @Content(
+                                    mediaType = MediaType.APPLICATION_JSON_VALUE,
+                                    schema = @Schema(implementation = ProblemJson.class))),
+                    @ApiResponse(
+                            responseCode = "500",
+                            description = "Service unavailable.",
+                            content =
+                            @Content(
+                                    mediaType = MediaType.APPLICATION_JSON_VALUE,
+                                    schema = @Schema(implementation = ProblemJson.class)))
+            })
+    @PostMapping(
+            value = "/organizations/{organizationfiscalcode}/transfers/iban/update",
+            produces = {"application/json"},
+            consumes = {"application/json"})
+    void updateTransferIbanMassive(
+            @Parameter(
+                    description = "Organization fiscal code, the fiscal code of the Organization.",
+                    required = true)
+            @PathVariable("organizationfiscalcode")
+            String organizationFiscalCode,
+            @Valid @RequestBody UpdateTransferIbanMassiveModel updateTransferIbanMassiveModel);
 }
