+++ src/main/java/it/gov/pagopa/debtposition/controller/payments/api/impl/PaymentsController.java
@@ -5,6 +5,7 @@ import it.gov.pagopa.debtposition.entity.PaymentOption;
 import it.gov.pagopa.debtposition.entity.Transfer;
 import it.gov.pagopa.debtposition.exception.AppError;
 import it.gov.pagopa.debtposition.exception.AppException;
+import it.gov.pagopa.debtposition.model.payments.UpdateTransferIbanMassiveModel;
 import it.gov.pagopa.debtposition.model.payments.PaymentOptionModel;
 import it.gov.pagopa.debtposition.model.payments.response.PaymentOptionModelResponse;
 import it.gov.pagopa.debtposition.model.payments.response.PaymentOptionWithDebtorInfoModelResponse;
@@ -142,5 +143,23 @@ public class PaymentsController implements IPaymentsController {
     }
     throw new AppException(
         AppError.PAYMENT_OPTION_NOTIFICATION_FEE_UPDATE_FAILED, organizationFiscalCode, iuv);
+  }
+
+  @Override
+  public void updateTransferIbanMassive(
+          String organizationFiscalCode,
+          @Valid UpdateTransferIbanMassiveModel updateTransferIbanMassiveModel) {
+    log.debug(
+            String.format(
+                    LOG_BASE_HEADER_INFO,
+                    "POST",
+                    "updateIbanMassive",
+                    String.format(
+                            "organizationFiscalCode= %s; oldIban= %s; newIban= %s",
+                            organizationFiscalCode,
+                            updateTransferIbanMassiveModel.getOldIban(),
+                            updateTransferIbanMassiveModel.getNewIban())));
+
+    paymentsService.updateTransferIbanMassive(organizationFiscalCode, updateTransferIbanMassiveModel.getOldIban(), updateTransferIbanMassiveModel.getNewIban());
   }
 }
